PID test 1 works Position control for the Left drive worked wonderfully when the PID controller was tuned (for off the carpet only). The system is not perfectly tuned, as it does not stay in position as required (possibly needs more D).
The rightEncoder took some time to get working, as we were mapping it to
the speed of the left encoder, as opposed to a position. We also
determined that the ReturnPIDInput returns a value between 0 and 1, so
we had to modify the command to make 1 full speed and 0 stopped.

This update so far has worked to make the robot drive straight. Off-carpet tests showed the encoders going to the same position at around the same speed, with some steady-state error in the right encoder.

Next steps involve making the input to the command be a percentage of full
speed (if it isn't already).