Fixed joystick mismatch/updated operator console
Joysticks were mismatched - "leftdriver" was matched up to the right side.
This was fixed in code instead of in electrical.
The operator console now has a "pacman" function that uses a timer to
flash the lights on the console on and off. It is run in the disabled
method, though we can call it from anywhere.
The potentiometer reads values, but they are tend to jump up and down.
This is likely because we are not calibrating it properly, though it might
be because we are calling the wrong joystick axis (it is mapped to more
than one).